Seperated harmonic analysis functions into their own file. Use harmonic_fit function in timeseries
